requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
pytz==2024.1
aiofiles==23.2.1

